Revert "Restructured project files"
