+++ application/frontend/wms/src/components/Main.jsx
@@ -1,17 +1,20 @@
 import React from "react";
-import { DashboardMetrics, DashboardStock, Header, Navbar } from "../../components"
+import Navbar from "./shared/Navbar";
+import Header from "./shared/Header";
+import DashboardStatsGrid from "./DashboardStatsGrid";
+import RecentProducts from "./RecentProducts";
 
-export default function Home() {
+export default function Main() {
   return (
     <div className="flex flex-row bg-neutral-100 h-screen w-screen overflow-hidden">
       <Navbar />
       <div className="flex-1">
         <Header />
         <div className="flex gap-4">
-          <DashboardMetrics />
+          <DashboardStatsGrid />
         </div>
         <div className="flex flex-row gap-4 m-4">
-          <DashboardStock />
+          <RecentProducts />
         </div>
       </div>
     </div>
