+++ application/frontend/wms/src/components/Products.jsx
@@ -1,0 +1,13 @@
+import React from "react";
+import Navbar from "./shared/Navbar";
+
+export default function Dashboard() {
+    return <div className="flex">
+        <div>
+            <Navbar />
+        </div>
+        <div>
+        <h1>Products page</h1>
+        </div>
+    </div>
+}
