+++ application/frontend/wms/src/components/Settings.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Navbar from "../../components/Navigation/Navbar";
+import Navbar from "./shared/Navbar";
 
 export default function Dashboard() {
   return (
