--- application/frontend/wms/src/components/index.jsx
@@ -1,4 +1,0 @@
-export { default as DashboardMetrics } from './Dashboard/DashboardMetrics'
-export { default as DashboardStock } from './Dashboard/DashboardStock'
-export { default as Header } from './Header/Header'
-export { default as Navbar } from './Navigation/Navbar'
