--- application/frontend/wms/src/routes/index.jsx
@@ -1,4 +1,0 @@
-export { default as Home } from './home/Home'
-export { default as Login } from './login/Login'
-export { default as ProductsList } from './products/ProductsList'
-export { default as Settings } from './settings/Settings'
